ACPI: Disable smart battery manager on Apple
Touching the smart battery manager at all on Apple hardware appears to
make it unhappy - unplugging the AC adapter triggers accesses that hang
the controller for several minutes. Quirk it out via DMI in order to
avoid this.  Compensate by changing battery presence if we fail to
communicate with the battery.

Signed-off-by: Matthew Garrett <matthew.garrett@nebula.com>